backend/pyproject.toml
[project]
name = "stockmaster-backend"
version = "0.1.0"
description = "StockMaster — Indian-market stock-insights backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.29",
    "gunicorn>=22",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5",
    "apscheduler>=3.10,<4",
    "firebase-admin>=6.5",
    "anthropic>=0.25",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx",
    "ruff>=0.4",
    "mypy>=1.10",
    "coverage>=7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008", "UP042"]  # B008: FastAPI Depends() pattern; UP042: StrEnum compat with SQLAlchemy

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["app/__pycache__/*", "app/db/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
